pyproject.toml
[tool.poetry]
name = "django-markdown-overview-demo"
version = "0.1.0"
description = ""
authors = ["Sqvall <masster1987@inbox.ru>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "3.10.2"
django-environ = "^0.9.0"
django-storages = "^1.13.1"
psycopg2-binary = "^2.9.3"
boto3 = "^1.24.70"
martor = "^1.6.14"
django-mdeditor = "^0.1.20"
django-markdownx = "^3.0.1"
django-simplemde = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
